[#9915] Moved Intent.java from action to request package
