+++ .gitignore
@@ -35,6 +35,9 @@ test-output/*
 *.iml
 out/
 
+# VS code
+.vscode/settings.json
+
 # JMeter
 jmeter.log
 src/e2e/lnp/*
