+++ build.gradle
@@ -157,8 +157,7 @@ task generateApiOutputFormat(type: cz.habarta.typescript.generator.gradle.Genera
 
 task generateApiRequestFormat(type: cz.habarta.typescript.generator.gradle.GenerateTask) {
     classPatterns = [
-        "teammates.ui.webapi.request.**",
-        "teammates.ui.webapi.action.Intent"
+        "teammates.ui.webapi.request.**"
     ]
     outputFile = "src/web/types/api-request.ts"
 }
