+++ src/main/java/teammates/ui/webapi/action/DeleteFeedbackResponseAction.java
@@ -13,6 +13,7 @@ import teammates.common.exception.EntityNotFoundException;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.exception.UnauthorizedAccessException;
 import teammates.common.util.Const;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * Delete a feedback response.
