+++ src/main/java/teammates/ui/webapi/action/GetFeedbackSessionAction.java
@@ -6,6 +6,7 @@ import teammates.common.datatransfer.attributes.StudentAttributes;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.util.Const;
 import teammates.ui.webapi.output.FeedbackSessionData;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * Get a feedback session.
