+++ src/main/java/teammates/ui/webapi/action/GetInstructorsAction.java
@@ -11,6 +11,7 @@ import teammates.common.exception.EntityNotFoundException;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.util.Const;
 import teammates.ui.webapi.output.InstructorsData;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * Get a list of instructors of a course.
