+++ src/main/java/teammates/ui/webapi/action/UpdateFeedbackResponseAction.java
@@ -16,6 +16,7 @@ import teammates.common.exception.UnauthorizedAccessException;
 import teammates.common.util.Const;
 import teammates.ui.webapi.output.FeedbackResponseData;
 import teammates.ui.webapi.request.FeedbackResponseUpdateRequest;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * Updates a feedback response.
