+++ src/test/java/teammates/test/cases/webapi/DeleteFeedbackResponseActionTest.java
@@ -13,8 +13,8 @@ import teammates.common.exception.EntityNotFoundException;
 import teammates.common.exception.InvalidHttpParameterException;
 import teammates.common.util.Const;
 import teammates.ui.webapi.action.DeleteFeedbackResponseAction;
-import teammates.ui.webapi.action.Intent;
 import teammates.ui.webapi.action.JsonResult;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * SUT: {@link DeleteFeedbackResponseAction}.
