+++ src/test/java/teammates/test/cases/webapi/GetFeedbackSessionActionTest.java
@@ -7,13 +7,13 @@ import teammates.common.datatransfer.attributes.FeedbackSessionAttributes;
 import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.util.Const;
 import teammates.ui.webapi.action.GetFeedbackSessionAction;
-import teammates.ui.webapi.action.Intent;
 import teammates.ui.webapi.action.JsonResult;
 import teammates.ui.webapi.output.FeedbackSessionData;
 import teammates.ui.webapi.output.FeedbackSessionPublishStatus;
 import teammates.ui.webapi.output.FeedbackSessionSubmissionStatus;
 import teammates.ui.webapi.output.ResponseVisibleSetting;
 import teammates.ui.webapi.output.SessionVisibleSetting;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * SUT: {@link GetFeedbackSessionAction}.
