+++ src/test/java/teammates/test/cases/webapi/UpdateFeedbackResponseActionTest.java
@@ -16,9 +16,9 @@ import teammates.common.datatransfer.questions.FeedbackQuestionType;
 import teammates.common.exception.InvalidHttpRequestBodyException;
 import teammates.common.util.Const;
 import teammates.test.driver.AssertHelper;
-import teammates.ui.webapi.action.Intent;
 import teammates.ui.webapi.action.UpdateFeedbackResponseAction;
 import teammates.ui.webapi.request.FeedbackResponseUpdateRequest;
+import teammates.ui.webapi.request.Intent;
 
 /**
  * SUT: {@link UpdateFeedbackResponseAction}.
